:lipstick: 최상단 페이지에서 사용할 Bottom Navigation 추가
