+++ lib/presentation/widget/desserbee_bottom_navigation.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:frontend_mobile/common/design_system/component/navigation_bar/navigation_bar.dart';
+import 'package:frontend_mobile/common/gen_asset/assets.gen.dart';
+import 'package:frontend_mobile/presentation/router/routes.dart';
+import 'package:go_router/go_router.dart';
+
+// 앱의 최상단 페이지에서 사용될 Bottom Navigation Bar
+class DesserbeeBottomNavigation extends StatelessWidget {
+  const DesserbeeBottomNavigation({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return CustomNavigationBar(
+      list: <NavigationBarType>[
+        NavigationBarType(
+          svg: Assets.icon.map.mapLine,
+          label: '지도',
+          onTap: () {
+            context.goNamed(AppRoutes.map.name);
+          },
+        ),
+      ],
+    );
+  }
+}
